# Conflicts:
#	ComeBackHome/src/main/java/TheEarthGuard/ComeBackHome/controller/CaseController.java
#	ComeBackHome/src/main/java/TheEarthGuard/ComeBackHome/repository/CaseRepository.java
#	ComeBackHome/src/main/java/TheEarthGuard/ComeBackHome/repository/JpaCaseRepository.java
#	ComeBackHome/src/main/java/TheEarthGuard/ComeBackHome/service/CaseService.java

+++ ComeBackHome/src/main/java/TheEarthGuard/ComeBackHome/controller/ReportController.java
@@ -1,15 +1,17 @@
 package TheEarthGuard.ComeBackHome.controller;
 
+import TheEarthGuard.ComeBackHome.domain.Case;
 import TheEarthGuard.ComeBackHome.domain.Report;
 import TheEarthGuard.ComeBackHome.domain.User;
-import TheEarthGuard.ComeBackHome.dto.ReportFormDto;
 import TheEarthGuard.ComeBackHome.dto.ReportPlaceInfoDto;
 import TheEarthGuard.ComeBackHome.dto.ReportRequestDto;
+import TheEarthGuard.ComeBackHome.dto.ReportResponseDto;
 import TheEarthGuard.ComeBackHome.security.CurrentUser;
 import TheEarthGuard.ComeBackHome.service.CaseService;
 import TheEarthGuard.ComeBackHome.service.ReportService;
 
 import java.util.List;
+import java.util.Optional;
 
 import TheEarthGuard.ComeBackHome.service.UserService;
 import lombok.extern.slf4j.Slf4j;
@@ -18,8 +20,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 @Slf4j
@@ -40,19 +42,28 @@ public class ReportController {
     }
 
     //처음 제보 등록할 때
-    @GetMapping(value = "/reports/new")
-    public String createForm(Model model) {
-        model.addAttribute("reportForm", new ReportFormDto());
+    @GetMapping(value = "/reports/new/{id}")
+    public String createForm(Model model, @PathVariable("id") Long id, @CurrentUser User user) throws IllegalAccessException {
+        ReportRequestDto reportDto = new ReportRequestDto();
+        if (user == null) {
+            return "redirect:/cases";
+        }
+        Optional<Case> caseDto = caseService.findCase(id);
+        reportDto.setCases(caseDto.get());
+        model.addAttribute("reportForm", reportDto);
+
         return "reports/createReportForm";
     }
 
     //실종위치 찍고나서 제보글 이어서 작성할 때
-    @PostMapping(value="/reports/new")
-    public String updateReportForm(@ModelAttribute ReportPlaceInfoDto reportPlaceInfoDto, @ModelAttribute("reportForm") ReportFormDto reportForm, HttpServletRequest request, Model model){
-
+    @PostMapping(value="/reports/new/{id}")
+    public String createFormPlace(@ModelAttribute ReportPlaceInfoDto reportPlaceInfoDto, @ModelAttribute("reportForm") ReportRequestDto reportForm,
+                                  @PathVariable("id") Long id, Model model){
         reportForm.setWitness_area(reportPlaceInfoDto.getWitness_area());
         reportForm.setWitness_lat(reportPlaceInfoDto.getWitness_lat());
         reportForm.setWitness_lng(reportPlaceInfoDto.getWitness_lng());
+        Optional<Case> caseDto = caseService.findCase(id);
+        reportForm.setCases(caseDto.get());
 
         model.addAttribute("reportForm", reportForm);
         return "reports/createReportForm";
@@ -60,26 +71,25 @@ public class ReportController {
 
 
     //제보 제출
-    @PostMapping(value = "/reports/new/submit")
-    public String createReport(@Valid @ModelAttribute ReportRequestDto form, @CurrentUser User user,Errors errors){
+    @PostMapping(value = "/reports/new/{id}/submit")
+    public String createReport(@Valid @ModelAttribute ReportRequestDto form, @PathVariable("id") Long id, @CurrentUser User user,Errors errors) throws Exception {
         if (errors.hasErrors()) {
             System.out.println("ERROR!!!!!!!!");
             return "/";
         }
 
         User currentUser = userService.findByEmail(user.getEmail());
-//        Case caseObj= caseService.findCases().get(1);
         if (currentUser != null) {
-            reportService.uploadReport(currentUser.getId(), 1L, form);
+            reportService.uploadReport(currentUser.getId(), id, form, form.getWitnessPics());
         }else{
             return "/users/login";
         }
 
-        return "redirect:/";
+        return "redirect:/cases";
 
     }
 
-    @GetMapping(value = "/reports")
+    @GetMapping(value = "/mypage/reports")
     public String reportList(Model model, @CurrentUser User user) {
         List<Report> reports = reportService.getReportsListByUser(user);
         model.addAttribute("reports", reports);
@@ -89,7 +99,9 @@ public class ReportController {
     //제보 상세보기
     @GetMapping(value = "/reports/detail/{id}")
     public String reportDetail(Model model, @PathVariable("id") Long id, @CurrentUser User user) {
-        model.addAttribute("report", reportService.getReportDetail(id));
+        Report reportDto = reportService.getReportDetail(id);
+        ReportResponseDto responseDto = new ReportResponseDto(reportDto, user);
+        model.addAttribute("report", responseDto);
         model.addAttribute("user", user);
         return "reports/reportDetail";
     }
@@ -97,55 +109,63 @@ public class ReportController {
     //제보 삭제하기
     @GetMapping(value = "/reports/delete/{id}")
     public String deleteReport(@PathVariable("id") Long id, @CurrentUser User user) {
-        reportService.deleteReport(id, user);
-        return "redirect:/reports";
+        Report report = reportService.getReportDetail(id);
+        if (user.getId() == report.getUser().getId()) {
+            reportService.deleteReport(id, user);
+            return "redirect:/cases";
+        }
+        return "redirect:/cases";
     }
 
     //제보 수정하기
     @GetMapping(value = "/reports/update/{id}")
-    public String updateReport(@PathVariable("id") Long id, @CurrentUser User user) {
-        return "redirect:/reports";
+    public String updateReportForm(Model model, @PathVariable("id") Long id, @CurrentUser User user) {
+        Report report = reportService.getReportDetail(id);
+        ReportResponseDto responseDto = new ReportResponseDto(report, user);
+        if (user.getId() == report.getUser().getId()) {
+            model.addAttribute("reportForm", responseDto);
+            return "/reports/reportUpdate";
+        }
+        return "redirect:/cases";
+    }
+
+    @PostMapping(value = "/reports/update/{id}")
+    public String updateReport(@Valid @ModelAttribute ReportRequestDto form, @PathVariable("id") Long id,
+                               @CurrentUser User user, Errors errors) throws IllegalAccessException {
+        if (errors.hasErrors()) {
+            log.info("error!!");
+            return "redirect:/cases";
+        }
+        Report report = reportService.getReportDetail(id);
+        reportService.updateReport(user.getId(), id, form);
+        return "redirect:/reports/detail/{id}";
+    }
+
+
+    //제보 신고하기
+    @PostMapping(value = "/reports/warn/{id}")
+    public String reportWarn(Model model, @PathVariable("id") Long id, @CurrentUser User user) {
+        reportService.warnReport(id, user);
+        log.info("신고하기 버튼 눌림");
+        model.addAttribute("report", reportService.getReportDetail(id));
+        model.addAttribute("user", user);
+        return "redirect:/reports/detail/{id}";
     }
 
 
     //지도로 목격위치 찍는 부분
-    @PostMapping(value="/reports/new/searchPlace")
-    public String searchPlace(@Valid @ModelAttribute ReportFormDto form, Model model, Errors errors) {
+    @PostMapping(value="/reports/new/{id}/searchPlace")
+    public String searchPlace(@ModelAttribute ReportRequestDto form, @RequestParam("witnessPics") MultipartFile file, Model model, @PathVariable("id") Long id, Errors errors) {
         if (errors.hasErrors()) {
             System.out.println("ERROR!!!!!!!!");
             return "/reports/createReportForm";
         }
+        Optional<Case> caseDto = caseService.findCase(id);
+        form.setCases(caseDto.get());
+        log.info("searchplace : "+ caseDto.get().getCaseId());
         model.addAttribute("reportForm", form);
         return "/reports/searchPlace";
     }
 
-
-//    //연동(최신)
-//    @GetMapping(value = "/cases/{cases_id}/report")
-//    public List<Report> getCaseReports(@PathVariable Long cases_id){
-//        Case cases = caseService.findOne(cases_id).get();
-//        return
-//
-//    }
-
-    @PostMapping(value="/cases/{cases_id}/report")
-    public String addReport(@PathVariable("cases_id") Long cases_id, @CurrentUser User user, @ModelAttribute ReportRequestDto reportRequestDto, Model model, Errors errors){
-        if(errors.hasErrors()){
-            System.out.println("Error!!");
-            return "/";
-        }
-
-//        String username=user.getUsername();
-//        Optional<Case> cases = caseService.findOne(cases_id);
-//        reportRequestDto.setUser(user);
-//        reportRequestDto.setCases(cases.get());
-
-        reportService.uploadReport(user.getId(), 1L, reportRequestDto);
-//        reportService.uploadReport(user.getId(), case_id, reportRequestDto);
-
-
-        return "reports/createReportForm";
-
-    }
 
 }
